requirements.txt
flask==3.0.2
python-docx==1.1.0
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.3
scikit-learn==1.4.1.post1
nltk==3.8.1
python-dotenv==1.0.1